requirements.txt
Django
djangorestframework
drf-yasg
django-cors-headers
django-otp
